requirements.txt
opencv-python-headless
numpy
pytesseract
ultralytics
websockets
psycopg2-binary
python-dotenv
